requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
asyncpy==1.2.0
attrs==25.4.0
click==8.3.1
cssselect==1.4.0
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.5
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
jmespath==1.1.0
lxml==6.0.2
multidict==6.7.1
packaging==26.0
parsel==1.11.0
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
w3lib==2.4.0
yarl==1.22.0
